==> 02_DO_NOT_RUN_hatebase_api.ipynb <==
## Hatebase api key and endpoint

#https://github.com/hatebase/Hatebase-API-Docs/blob/master/current/v4-2/get_vocabulary.md
import pandas as pd
import json
import requests
import numpy as np
api_key = 'KAKHM8ADQdCKWNPjAFDGRYUXFNNzEVXW'  #api key for hatebase
api_url = 'https://api.hatebase.org/4-2/authenticate'  #authentication endpoint

## Authenticate with hatebase

data = {'api_key': api_key}
response = requests.post(url=api_url, data=data)  #call the api to authenticate and receive a token
resp = response.json()
token = resp['result']['token'] #the token for further queries

## Call the Hatebase api to find the number of results

api_url = 'https://api.hatebase.org/4-2/get_vocabulary' #endpoint for hate terms
data = {'token': token, 'language':'ENG'}
response = requests.post(url=api_url, data=data)
resp = response.json()
pages = resp['number_of_pages'] #count the number of pages of hate terms

## Loop through the results to create dataframe of terms

terms = [] #empty series
ave_offense = [] #empty series
for p in range(1,resp['number_of_pages']+1):
    data = {'token': token, 'language':'ENG', 'page':p}
    response = requests.post(url=api_url, data=data) #call each page
    resp = response.json()
    for t in range(0,len(resp['result'])): #get each term on a page
        terms = np.append(terms,resp['result'][t]['term']) #append to the series of terms
        ave_offense = np.append(ave_offense,resp['result'][t]['average_offensiveness']) #append to the series of terms
hate_terms = pd.DataFrame({'hate_terms':terms,
                          'ave_offense':ave_offense}) #create dataframe of terms
hate_terms.to_pickle('pickle_files/hate_terms.pickle')

hate_terms.to_csv('csv_excel/hate_terms2.csv')
