==> 04_data_cleansing.ipynb <==
## Common Imports

import pandas as pd
import numpy as np
from datetime import datetime
import re

## Remove urls, usernames, hashtags, and strip whitespace - all API tweets

current_time = datetime.now()

existing_data_formatted = pd.read_pickle('pickle_files/existing_data.pickle')

for i in existing_data_formatted.index:
    #print(i)
    existing_data_formatted.loc[i,'Tweet'] =re.sub("@[A-Za-z0-9_/:().]+",  "", existing_data_formatted.loc[i,'Tweet'])
    existing_data_formatted.loc[i,'Tweet'] =re.sub("http[A-Za-z0-9_/:().]+",  "", existing_data_formatted.loc[i,'Tweet'])
    existing_data_formatted.loc[i,'Tweet'] =re.sub("#[A-Za-z0-9_/:().]+",  "", existing_data_formatted.loc[i,'Tweet'])
existing_data_formatted.Tweet = existing_data_formatted.Tweet.str.strip()
existing_data_formatted = existing_data_formatted[existing_data_formatted['Tweet'].apply(lambda x: len(x) > 3)]
existing_data_formatted = existing_data_formatted.drop_duplicates(subset=['Tweet ID'], keep='first')
print("Formatting took time ", datetime.now() - current_time)

existing_data_formatted.to_pickle('pickle_files/api_formatted.pickle')

## Remove urls, usernames, hashtags, and strip whitespace - benign tweets

current_time = datetime.now()

existing_clean_formatted = pd.read_pickle('pickle_files/existing_clean.pickle')

for i in existing_clean_formatted.index:
    #print(i)
    existing_clean_formatted.loc[i,'Tweet'] =re.sub("@[A-Za-z0-9_/:().]+",  "", existing_clean_formatted.loc[i,'Tweet'])
    existing_clean_formatted.loc[i,'Tweet'] =re.sub("http[A-Za-z0-9_/:().]+",  "", existing_clean_formatted.loc[i,'Tweet'])
    existing_clean_formatted.loc[i,'Tweet'] =re.sub("#[A-Za-z0-9_/:().]+",  "", existing_clean_formatted.loc[i,'Tweet'])
existing_clean_formatted.Tweet = existing_clean_formatted.Tweet.str.strip()
existing_clean_formatted = existing_clean_formatted[existing_clean_formatted['Tweet'].apply(lambda x: len(x) > 3)]
existing_clean_formatted = existing_clean_formatted.drop_duplicates(subset=['Tweet ID'], keep='first')
print("Formatting took time ", datetime.now() - current_time)

existing_clean_formatted.to_pickle('pickle_files/benign_formatted.pickle')

## Remove urls, usernames, hashtags, and strip whitespace - offensive tweets

current_time = datetime.now()

existing_hate_formatted = pd.read_pickle('pickle_files/existing_hate.pickle')
existing_hate_formatted = existing_hate_formatted.reset_index(drop=True) #reset the index

for i in existing_hate_formatted.index:
    #print(i)
    existing_hate_formatted.loc[i,'Tweet'] =re.sub("@[A-Za-z0-9_/:().]+",  "", existing_hate_formatted.loc[i,'Tweet'])
    existing_hate_formatted.loc[i,'Tweet'] =re.sub("http[A-Za-z0-9_/:().]+",  "", existing_hate_formatted.loc[i,'Tweet'])
    existing_hate_formatted.loc[i,'Tweet'] =re.sub("#[A-Za-z0-9_/:().]+",  "", existing_hate_formatted.loc[i,'Tweet'])
    existing_hate_formatted.loc[i,'Tweet'] =re.sub("RT ",  "", existing_hate_formatted.loc[i,'Tweet'])
existing_hate_formatted.Tweet = existing_hate_formatted.Tweet.str.strip()
existing_hate_formatted = existing_hate_formatted[existing_hate_formatted['Tweet'].apply(lambda x: len(x) > 3)]
print("Formatting took time ", datetime.now() - current_time)

existing_hate_formatted.to_pickle('pickle_files/offensive_formatted.pickle')

print(existing_data_formatted.shape)
print(existing_clean_formatted.shape)
print(existing_hate_formatted.shape)
print(existing_hate_formatted.shape[0]+existing_clean_formatted.shape[0])
