==> 06_train_eval_split.ipynb <==
## Common imports

import pandas as pd
import numpy as np

## Train test split

offensive_formatted = pd.read_pickle('pickle_files/offensive_formatted.pickle')
benign_formatted = pd.read_pickle('pickle_files/benign_minusbj_formatted.pickle')

offensive_formatted['class'] = 1
benign_formatted['class'] = 0
df = offensive_formatted.append(benign_formatted)

from sklearn.utils import shuffle
df = shuffle(df, random_state=42)

from sklearn.model_selection import train_test_split
train_data_formatted,eval_data_formatted = train_test_split(df, test_size=0.2, random_state=42)

train_data_formatted.to_pickle('pickle_files/train_data_formatted.pickle')
eval_data_formatted.to_pickle('pickle_files/eval_data_formatted.pickle')

## Dataset sizes

api = pd.read_pickle('pickle_files/api_formatted.pickle')
benign = pd.read_pickle('pickle_files/benign_minusbj_formatted.pickle')
offensive = pd.read_pickle('pickle_files/offensive_formatted.pickle')
eval_data = pd.read_pickle('pickle_files/eval_data_formatted.pickle')
train_data = pd.read_pickle('pickle_files/train_data_formatted.pickle')
print("api shape "+ str(api.shape))
print("benign shape "+ str(benign.shape))
print("offensive shape "+ str(offensive.shape))
print("eval_data "+ str(eval_data.shape))
print("eval_data class 1 "+str(eval_data[eval_data['class']==1].shape))
print("eval_data class 0 "+str(eval_data[eval_data['class']==0].shape))
print("train_data "+ str(train_data.shape))
print("train_data class 1 "+str(train_data[train_data['class']==1].shape))
print("train_data class 0 "+str(train_data[train_data['class']==0].shape))
