==> 07e_counts.ipynb <==
## Common Imports

import pandas as pd
import numpy as np
import nltk
from nltk.tokenize import word_tokenize
nltk.download('punkt')
from nltk.corpus import stopwords
import re
import glob

## Load the data

train_data = pd.read_pickle('pickle_files/eval_data_formatted.pickle')
eval_data = pd.read_pickle('pickle_files/train_data_formatted.pickle')
all_data = train_data.append(eval_data)
all_data['Tweet ID'] = all_data['Tweet ID'].astype(str) #change the ID to str to avoid potential issues during aggregation
all_data = all_data.reset_index(drop=True)

## Load Stopwords

stop = stopwords.words('english') #get stopwords from NLTK
keep = ['not'] #Waseem/Hovy did not use "not" as a stopword
stop = [word for word in stop if word not in keep] #Waseem/Hovy did not use "not" as a stopword

## Lowercase the data

all_data['Tweet_original'] = all_data.Tweet.copy() #keep a copy of the original tweet text
all_data['Tweet'] = all_data['Tweet'].str.lower() #lowercase the text 

## Remove punctuation, usernames, hashtags, URLs

all_data['Tweet'] = all_data['Tweet'].fillna('')
p = re.compile(r'[^\w\s]+')
all_data['Tweet'] = [p.sub('', x) for x in all_data['Tweet'].tolist()] #remove the punctuation
for i in all_data.index:
    #print(i)
    #all_data.loc[i,'Tweet'] =re.sub('[^A-Za-z0-9]+',"",all_data.loc[i,'Tweet'])
    all_data.loc[i,'Tweet'] =re.sub("@[A-Za-z0-9_/:().]+",  "", all_data.loc[i,'Tweet'])
    all_data.loc[i,'Tweet'] =re.sub("http[A-Za-z0-9_/:().]+",  "", all_data.loc[i,'Tweet'])
    all_data.loc[i,'Tweet'] =re.sub("#[A-Za-z0-9_/:().]+",  "", all_data.loc[i,'Tweet'])
print(len(np.unique(all_data['Tweet ID']))) #for convenience

## Remove stopwords

all_data['Tweet'] = all_data['Tweet'].apply(lambda x: ' '.join([word for word in x.split() if word not in (stop)])) #remove stopwords
print(len(np.unique(all_data['Tweet ID']))) #for convenience

## Tokenize

all_data['Tweet'] = all_data['Tweet'].apply(word_tokenize) #tokenize the text
print(len(np.unique(all_data['Tweet ID']))) #for convenience

## Initialise arrays to hold results

tweet_ids = []
feature = []
value = []

## Count of tokens

for i in all_data.index:
    tweet_ids = np.append(tweet_ids,all_data['Tweet ID'][i])
    feature = np.append(feature,'token_count')
    value = np.append(value,len(all_data.Tweet[i]))
print(len(value))
print(len(feature))
print(len(tweet_ids))    

## Ave chars per token

for i in all_data.index:
    tweet_ids = np.append(tweet_ids,all_data['Tweet ID'][i])
    feature = np.append(feature,'ave_chars_token') 
    chars = 0
    for w in all_data.Tweet[i]:
        chars = chars + len(w)
    
    if len(all_data.Tweet[i])==0:
        value = np.append(value,0)
    else:
        value = np.append(value,chars/len(all_data.Tweet[i]))
print(len(value))
print(len(feature))
print(len(tweet_ids))

## Count of URLs

all_data.Web = all_data.Web.fillna('')
tweet_ids = np.append(tweet_ids,all_data['Tweet ID'])
feature = np.append(feature,np.repeat('url_count',len(all_data.index)))
value = np.append(value,all_data.Web.str.count('http'))
print(len(value))
print(len(feature))
print(len(tweet_ids))

## Count of Hashtags

all_data.Hashtags = all_data.Hashtags.fillna('')
for i in all_data.index:
    tweet_ids = np.append(tweet_ids,all_data['Tweet ID'][i])
    feature = np.append(feature,'hashtag_count')
    if (len(all_data.Hashtags[i])==0):
        value = np.append(value,all_data.Hashtags[i].count(' '))
    else:
        value = np.append(value,all_data.Hashtags[i].count(' ')+1)
    

## Count of Mentions

all_data.Mentions = all_data.Mentions.fillna('')
for i in all_data.index:
    tweet_ids = np.append(tweet_ids,all_data['Tweet ID'][i])
    feature = np.append(feature,'mention_count')
    if (len(all_data.Mentions[i])==0):
        value = np.append(value,all_data.Mentions[i].count(' '))
    else:
        value = np.append(value,all_data.Mentions[i].count(' ')+1)
print(len(value))
print(len(feature))
print(len(tweet_ids))

## Count of Caps

for i in all_data.index:
    tweet_ids = np.append(tweet_ids,all_data['Tweet ID'][i])
    feature = np.append(feature,'caps_count')
    value = np.append(value,sum(1 for c in all_data.Tweet_original[i] if c.isupper()))
    
print(len(value))
print(len(feature))
print(len(tweet_ids))

## Count of Characters

for i in all_data.index:
    tweet_ids = np.append(tweet_ids,all_data['Tweet ID'][i])
    feature = np.append(feature,'character_count')
    value = np.append(value,len(all_data.Tweet_original[i]))

print(len(value))
print(len(feature))
print(len(tweet_ids))

## Count of followers

tweet_ids = np.append(tweet_ids,all_data['Tweet ID'])
feature = np.append(feature,np.repeat('followers_count',len(all_data.index)))
value = np.append(value,all_data['Number of Followers'])

print(len(value))
print(len(feature))
print(len(tweet_ids))

## Count of Following

tweet_ids = np.append(tweet_ids,all_data['Tweet ID'])
feature = np.append(feature,np.repeat('following_count',len(all_data.index)))
value = np.append(value,all_data['Number Following'])

print(len(value))
print(len(feature))
print(len(tweet_ids))

## Count of Posted Tweets

tweet_ids = np.append(tweet_ids,all_data['Tweet ID'])
feature = np.append(feature,np.repeat('posted_tweets_count',len(all_data.index)))
value = np.append(value,all_data['Number of Tweets'])

print(len(value))
print(len(feature))
print(len(tweet_ids))

## Count of Modals

modals = ['can', 'could', 'may', 'might', 'must', 'will', 'would', 'should']
for i in all_data.index:
    tweet_ids = np.append(tweet_ids,all_data['Tweet ID'][i])
    feature = np.append(feature,'modals_count')
    
    count = 0
    for m in modals:
        for t in all_data.Tweet[i]:
            if m==t:
                count = count+1    
    
    value = np.append(value,count)    

print(len(value))
print(len(feature))
print(len(tweet_ids))

## Count of tokens with non-alpha characters in the middle (masked)

rx = re.compile("^[a-zA-ZáéíóúüÁÉÍÓÚÜ]+[0-9]+[a-zA-ZáéíóúüÁÉÍÓÚÜ]+$") #regex for alpha - non-alpha - alpha
for i in all_data.index:
    tweet_ids = np.append(tweet_ids,all_data['Tweet ID'][i])
    feature = np.append(feature,'masked_count')
    
    count = 0
    for w in all_data.Tweet[i]:
        count = count + len(rx.findall(w))
    
    value = np.append(value,count) 
    
print(len(value))
print(len(feature))
print(len(tweet_ids))

## Count of 1 character tokens

for i in all_data.index:
    tweet_ids = np.append(tweet_ids,all_data['Tweet ID'][i])
    feature = np.append(feature,'1char_token_count')
    
    count=0
    for t in all_data.Tweet[i]:
        if len(t)==1:
            if str.isalpha(t):
                count = count+1
    value = np.append(value,count)

print(len(value))
print(len(feature))
print(len(tweet_ids))

## Count of quotations

rx = re.compile('\"(.+?)\"') #regex for double quoted text
for i in all_data.index:
    tweet_ids = np.append(tweet_ids,all_data['Tweet ID'][i])
    feature = np.append(feature,'quotes_count')
    value = np.append(value,len(rx.findall(all_data.Tweet_original[i]))) 

print(len(value))
print(len(feature))
print(len(tweet_ids))

## Count of punctuation

import string
count = lambda l1, l2: len(list(filter(lambda c: c in l2, l1)))
for i in all_data.index:
    tweet_ids = np.append(tweet_ids,all_data['Tweet ID'][i])
    feature = np.append(feature,'punctuation_count')
    value = np.append(value,count(all_data.Tweet_original[i], string.punctuation)) 

print(len(value))
print(len(feature))
print(len(tweet_ids))

## Sentiment

from nltk.stem import PorterStemmer
ps = PorterStemmer()
all_data['tweet_sent'] = all_data['Tweet']
for i in all_data.index:
    for w in range(0,len(all_data.tweet_sent[i])):
        all_data.tweet_sent[i][w] = ps.stem(all_data.tweet_sent[i][w])
        
from nltk.sentiment.vader import SentimentIntensityAnalyzer
nltk.download('vader_lexicon')
sid = SentimentIntensityAnalyzer()
all_data['tweet_sent_str'] = all_data.tweet_sent.str.join(" ")

tweet_ids = np.append(tweet_ids,all_data['Tweet ID'])
feature = np.append(feature,np.repeat('neg_sent',len(all_data.index)))
value = np.append(value,all_data['tweet_sent_str'].apply(lambda x: sid.polarity_scores(x)['neg'])) 

tweet_ids = np.append(tweet_ids,all_data['Tweet ID'])
feature = np.append(feature,np.repeat('neu_sent',len(all_data.index)))
value = np.append(value,all_data['tweet_sent_str'].apply(lambda x: sid.polarity_scores(x)['neu'])) 

tweet_ids = np.append(tweet_ids,all_data['Tweet ID'])
feature = np.append(feature,np.repeat('pos_sent',len(all_data.index)))
value = np.append(value,all_data['tweet_sent_str'].apply(lambda x: sid.polarity_scores(x)['pos'])) 

print(len(value))
print(len(feature))
print(len(tweet_ids))

df_f = pd.DataFrame({'tweet_ids':tweet_ids,
                        'feature':feature,
                        'value':value})
df_f.to_pickle('features/count_features/count_features.pickle')
df_f_pivot = df_f.pivot_table(index=['tweet_ids'], columns='feature', values='value').reset_index() #pivot
df_f_pivot.shape

df_f_pivot.iloc[20:30,:]
