==> 25_COLAB_github_save.ipynb <==
## Common Imports

import tensorflow as tf
import pandas as pd

## Authorise access to Google Cloud Storage

# Set the output directory for saving model file
# Optionally, set a GCP bucket location

OUTPUT_DIR = ''#@param {type:"string"}
#@markdown Whether or not to clear/delete the directory and create a new one
DO_DELETE = False #@param {type:"boolean"}
#@markdown Set USE_BUCKET and BUCKET if you want to (optionally) store model output on GCP bucket.
USE_BUCKET = True #@param {type:"boolean"}
BUCKET = 'dissertation_bucket' #@param {type:"string"}

if USE_BUCKET:
  OUTPUT_DIR = 'gs://{}/{}'.format(BUCKET, OUTPUT_DIR)
  from google.colab import auth
  auth.authenticate_user()

if DO_DELETE:
  try:
    tf.gfile.DeleteRecursively(OUTPUT_DIR)
  except:
    # Doesn't matter if the directory didn't exist
    pass
tf.gfile.MakeDirs(OUTPUT_DIR)
print('***** Model output directory: {} *****'.format(OUTPUT_DIR))

## Collate the metrics from the embeddings, CNN, and BERT experiments

#******fasttext 300 CNN******
!gsutil cp gs://dissertation_bucket/fasttext300_cnn_files/prediction_summary.pickle /tmp/
preds_per_tweet = pd.read_pickle('/tmp/prediction_summary.pickle')
print(preds_per_tweet.shape)

!gsutil cp gs://dissertation_bucket/fasttext300_cnn_files/metrics.csv /tmp/
metrics = pd.read_csv('/tmp/metrics.csv')


#******glove 300 CNN******
!gsutil cp gs://dissertation_bucket/glove300_840_cnn_files/prediction_summary.pickle /tmp/
df = pd.read_pickle('/tmp/prediction_summary.pickle')
preds_per_tweet = preds_per_tweet.append(df)
print(df.shape)
print(preds_per_tweet.shape)

!gsutil cp gs://dissertation_bucket/glove300_840_cnn_files/metrics.csv /tmp/
df = pd.read_csv('/tmp/metrics.csv')
metrics = metrics.append(df)


#******w2v CNN******
!gsutil cp gs://dissertation_bucket/w2v_cnn_files/prediction_summary.pickle /tmp/
df = pd.read_pickle('/tmp/prediction_summary.pickle')
preds_per_tweet = preds_per_tweet.append(df)
print(df.shape)
print(preds_per_tweet.shape)

!gsutil cp gs://dissertation_bucket/w2v_cnn_files/metrics.csv /tmp/
df = pd.read_csv('/tmp/metrics.csv')
metrics = metrics.append(df)


#******w2v RNN******
!gsutil cp gs://dissertation_bucket/w2v_rnn_files/prediction_summary.pickle /tmp/
df = pd.read_pickle('/tmp/prediction_summary.pickle')
preds_per_tweet = preds_per_tweet.append(df)
print(df.shape)
print(preds_per_tweet.shape)

!gsutil cp gs://dissertation_bucket/w2v_rnn_files/metrics.csv /tmp/
df = pd.read_csv('/tmp/metrics.csv')
metrics = metrics.append(df)


#******glove RNN******
!gsutil cp gs://dissertation_bucket/glove_rnn_files/prediction_summary.pickle /tmp/
df = pd.read_pickle('/tmp/prediction_summary.pickle')
preds_per_tweet = preds_per_tweet.append(df)
print(df.shape)
print(preds_per_tweet.shape)

!gsutil cp gs://dissertation_bucket/glove_rnn_files/metrics.csv /tmp/
df = pd.read_csv('/tmp/metrics.csv')
metrics = metrics.append(df)


#******fasttext RNN******
!gsutil cp gs://dissertation_bucket/fasttext_rnn_files/prediction_summary.pickle /tmp/
df = pd.read_pickle('/tmp/prediction_summary.pickle')
preds_per_tweet = preds_per_tweet.append(df)
print(df.shape)
print(preds_per_tweet.shape)

!gsutil cp gs://dissertation_bucket/fasttext_rnn_files/metrics.csv /tmp/
df = pd.read_csv('/tmp/metrics.csv')
metrics = metrics.append(df)


#******BERT LARGE******
!gsutil cp gs://dissertation_bucket/bert_large_files/prediction_summary.pickle /tmp/
df = pd.read_pickle('/tmp/prediction_summary.pickle')
preds_per_tweet = preds_per_tweet.append(df)
print(df.shape)
print(preds_per_tweet.shape)

!gsutil cp gs://dissertation_bucket/bert_large_files/metrics.csv /tmp/
df = pd.read_csv('/tmp/metrics.csv')
metrics = metrics.append(df)

#******BERT BASE******
!gsutil cp gs://dissertation_bucket/bert_base_files/prediction_summary.pickle /tmp/
df = pd.read_pickle('/tmp/prediction_summary.pickle')
preds_per_tweet = preds_per_tweet.append(df)
print(df.shape)
print(preds_per_tweet.shape)

!gsutil cp gs://dissertation_bucket/bert_base_files/metrics.csv /tmp/
df = pd.read_csv('/tmp/metrics.csv')
metrics = metrics.append(df)


#******classify_embeddings_fasttext******
!gsutil cp gs://dissertation_bucket/classify_embeddings_fasttext/df.pickle /tmp/
df = pd.read_pickle('/tmp/df.pickle')
preds_per_tweet = preds_per_tweet.append(df)
print(df.shape)
print(preds_per_tweet.shape)

!gsutil cp gs://dissertation_bucket/classify_embeddings_fasttext/classifications.pickle /tmp/
df = pd.read_pickle('/tmp/classifications.pickle')
metrics = metrics.append(df)


#******classify_embeddings_glove******
!gsutil cp gs://dissertation_bucket/classify_embeddings_glove/df.pickle /tmp/
df = pd.read_pickle('/tmp/df.pickle')
preds_per_tweet = preds_per_tweet.append(df)
print(df.shape)
print(preds_per_tweet.shape)

!gsutil cp gs://dissertation_bucket/classify_embeddings_glove/classifications.pickle /tmp/
df = pd.read_pickle('/tmp/classifications.pickle')
metrics = metrics.append(df)


#******classify_embeddings_glove_twitter200******
!gsutil cp gs://dissertation_bucket/classify_embeddings_glove_twitter200/df.pickle /tmp/
df = pd.read_pickle('/tmp/df.pickle')
preds_per_tweet = preds_per_tweet.append(df)
print(df.shape)
print(preds_per_tweet.shape)

!gsutil cp gs://dissertation_bucket/classify_embeddings_glove_twitter200/classifications.pickle /tmp/
df = pd.read_pickle('/tmp/classifications.pickle')
metrics = metrics.append(df)


#******classify_embeddings_w2v******
!gsutil cp gs://dissertation_bucket/classify_embeddings_w2v/df.pickle /tmp/
df = pd.read_pickle('/tmp/df.pickle')
preds_per_tweet = preds_per_tweet.append(df)
print(df.shape)
print(preds_per_tweet.shape)

!gsutil cp gs://dissertation_bucket/classify_embeddings_w2v/classifications.pickle /tmp/
df = pd.read_pickle('/tmp/classifications.pickle')
metrics = metrics.append(df)

#******model comparisons******
!gsutil cp gs://dissertation_bucket/bert_large_assess_files/metrics.csv /tmp/
compare_bert = pd.read_csv('/tmp/metrics.csv')

!gsutil cp gs://dissertation_bucket/glove_rnn_assess_files/metrics.csv /tmp/
compare_rnn = pd.read_csv('/tmp/metrics.csv')
compare = compare_bert.append(compare_rnn)




## Pickle the predictions per tweet and the metrics for each experiment

#pickle
metrics.to_pickle('/tmp/metrics_colab.pickle')
preds_per_tweet.to_pickle('/tmp/preds_per_tweet_colab.pickle')
compare.to_pickle('/tmp/compare.pickle')

## Clone the github repository

!git clone https://github.com/ScottJK-20190706/tweet_classifier


## Push the files to github

#copy files to the repo
!cp /tmp/preds_per_tweet_colab.pickle tweet_classifier/results
!cp /tmp/metrics_colab.pickle tweet_classifier/results
!cp /tmp/compare.pickle tweet_classifier/results

#change directory to the repo
import os
os.chdir('tweet_classifier/results')

#set username and password
!git config --global user.email "scott.kilgariff@gmail.com" #
!git config --global user.name "ScottJK-20190706"

#add
!git add preds_per_tweet_colab.pickle
!git add metrics_colab.pickle
!git add compare.pickle

#commit
!git commit --message="Add predictions per tweet from colab"
!git commit --message="Add metrics from colab"
!git commit --message="Add model comaprison from colab"

#push
!git remote set-url origin https://ScottJK-20190706:K1l64r1ff@github.com/ScottJK-20190706/tweet_classifier.git
!git push origin
